refactor: rename SuccessModal component
