+++ src/routes/Schedule/style.css
@@ -68,7 +68,7 @@
 
 .btn-sched {
     padding: 15px 30px;
-    background-color: var(--btn-bg);
+    background-color: var(--bg-blue);
     color: white;
     border-radius: 6px;
     border: none;
